+++ patching/plugin-commands-patching/test/patch.test.ts
@@ -1298,6 +1298,7 @@ describe('patch with custom modules-dir and virtual-store-dir', () => {
       selectedProjectsGraph,
       workspaceDir: customModulesDirFixture,
       saveLockfile: true,
+      confirmModulesPurge: false,
     })
     const output = await patch.handler(defaultPatchOption, ['is-positive@1'])
     const patchDir = getPatchDirFromPatchOutput(output)
@@ -1320,6 +1321,7 @@ describe('patch with custom modules-dir and virtual-store-dir', () => {
       selectedProjectsGraph,
       lockfileDir: customModulesDirFixture,
       workspaceDir: customModulesDirFixture,
+      confirmModulesPurge: false,
     }, [patchDir])
     expect(fs.readFileSync(path.join(customModulesDirFixture, 'packages/bar/fake_modules/is-positive/index.js'), 'utf8')).toContain('// test patching')
   })