+++ config/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/config
 
+## 1003.1.1
+
+### Patch Changes
+
+- 8d175c0: Use `pnpm_config_` env variables instead of `npm_config_` [#9571](https://github.com/pnpm/pnpm/pull/9571).
+  - @pnpm/pnpmfile@1001.2.2
+
 ## 1003.1.0
 
 ### Minor Changes